Update controller-runtime to v0.14.4
controller-runtime deprecated the envtest/printer package in v0.14.0. They don't
needed anymore since they upgraded to ginko v2. In order to keep the
scope contained, I just copied the code we needed, it's quite simple.
This allow us to maintain the same version of ginko while bumping
controller-runtime.

The fake client provided by controller-runtime was updated to allow
for field selectors in List operations. This requires to define the
index prior to its use. Since the Job controller uses
".metadata.controller" for a List query, we now set this index in the
tests client.
https://github.com/kubernetes-sigs/controller-runtime/pull/2025

Signed-off-by: Guillermo Gaston <gaslor@amazon.com>