Update controller logging, fix session leak:
This add logging to show metadata about all
bmc calls. Refactored logging in all controllers to
pull the logger from ctx. This is because the controller
runtime enriches the logs with a lot of detail that we
can take advantage of.

Removed the zap logger in favor of zerolog because its
of its smaller dependency footprint and so that logs go
to stdout instead of stderr by default.

Fixed a session leak in the task controller. This was happening
because the defer to close the session wasn't always being called.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>