Update linting version, fix linting issues:
The way the linter was run the exit code was
being swallowed up and always return 0. This
meant that even with linting issues could CI
was passing. This fixes that.

The new version of the linter found some unused
functions, so they were removed. The other changes
were to resolve linting issues that were found.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>