Update bmclib version:
Fixes a nil pointer panic when running in
environments (container images like scratch) that don't
have the "sum" binary. In bmclib the "sum" binary
is the supermicro update manager, but in most OS
distros "sum" is a command for
"checksum and count the blocks in a file". This means
that we didn't see the panic because we are running
in Alpine that have a "sum" binary.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>